Resolve "Add player health and balance and display on screen"

Closes #8

See merge request mjc/gameboygamedev!5-e 
Original/build commit ID: 4aa6bb1d11f6cbd7df57ed4a722d07880d91c240

+++ game_state.c
@@ -26,7 +26,7 @@ typedef struct {
     // Current building
     UINT8 current_building;
     // In-game days passed. Might need to be int (not required for limited game plays).
-    UINT8 days_passed;
+    unsigned int days_passed;
     
     // Current hour of day (out of 12).
     UINT8 hour;
@@ -38,5 +38,5 @@ typedef struct {
     UINT8 hp;
     
     // Money the user has.
-    unsigned long balance;
+    unsigned int balance;
 } game_state_t;
